Added support for non-tab 'li' elements.
Potential solution to issue #33. Optionaly use 'has:' jQuery selector
to isolate relevant 'li' elements. In order not to increase the version
of jQuery required, added logic to test for the support of the selector
which is only available in version 1.1.4 and up.  Testing for support
of a jQuery selector is not straight forward, so my implementation is a
bit hacky, but simple:

There doesn't appear to be a good way to test if the currently loaded
version of jQuery supports a particular selector. Nor is there a simple
way to determine that the version of jQuery loaded is at least at 'x'
(in this case 1.1.4). If we don't want to increase the required version
of jQuery for the plug-in, we need to test for this though. So instead
I test for a function, 'slice', that was added in version 1.1.4 as
well. This is somewhat fragile if a later version depricates 'slice',
but seems like the best approach for now if backward compatibility is
important.